DatasetPlotting/requirements.txt
# --- FILE: requirements.txt ---
numpy
pandas
scikit-learn
umap-learn
matplotlib
psycopg2-binary
tqdm
